+++ rapdev_zoom/datadog_checks/rapdev_zoom/__about__.py
@@ -1,3 +1,3 @@
 # (C) RapDev, Inc. 2020-present
 # All rights reserved
-__version__ = '2.0.1'
+__version__ = '3.0.0'