+++ rapdev_zoom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG - rapdev_zoom
 
+## 3.0.0 / 2021-02-17
+
+* [Removed] The registered users count metric.
+* [Added] The active users counts tagged by account type (e.g. basic, licensed, etc...).
+* [Removed] The user connection type tag.
+
 ## 2.0.1 / 2021-02-05
 
 * [Modified] Audio/Video Widget Titles to align with Sending/Receiving metrics, rather than input/output
@@ -11,4 +17,3 @@
 * [Added] added CPU Metric ingestion.
 * [Added] added users_to_track flag that allows you to specify a list of users for which the agent will only send metrics for.
 
-